Rename index.js to app.js and change directory to root
