+++ package.json
@@ -2,14 +2,14 @@
   "name": "questioner",
   "version": "1.0.0",
   "description": "Crowd-source questions for a scheduled meetup",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "start-local": "nodemon --exec babel-node server/index.js",
-    "start": "npm run build && node ./dist/index.js",
+    "start-local": "nodemon --exec babel-node ./app.js",
+    "start": "npm run build && node ./dist/app.js",
     "build": "npm run clean-build && npm run babel-build",
-    "babel-build": "babel ./server -d dist",
+    "babel-build": "babel ./ -d dist --ignore node_modules",
     "clean-build": "rm -rf dist",
-    "serve": "SET NODE_ENV=dev && node ./dist/index.js",
+    "serve": "SET NODE_ENV=dev && node ./dist/app.js",
     "test": "mocha ./test --compilers js:@babel/register --recursive --exit",
     "test-coverage": "cross-env NODE_ENV=test nyc mocha --exit",
     "report-coverage": "nyc report --reporter=text-lcov | coveralls"
