+++ test/api.v1/rsvps.spec.js
@@ -1,6 +1,6 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../../server';
+import server from '../../app';
 
 chai.use(chaiHttp);
